+++ src/main/java/com/advanced/app/log/v2/OrderControllerV2.java
@@ -1,7 +1,7 @@
 package com.advanced.app.log.v2;
 
 import com.advanced.trace.TraceStatus;
-import com.advanced.trace.hellotrace.HelloTraceV2;
+import com.advanced.app.log.hellotrace.HelloTraceV2;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
